chore: merge bus logic into arraial page
